requirements.txt
# For build a website
django>3.2,<3.3
# Have connection between environment and site
python-decouple
# setting up restframework
djangorestframework==3.14
djangorestframework-simplejwt
django-filter==2.4
# Swagger
drf-yasg[validation]
# Sending html form email
django-mail-templated
# Show images
Pillow
# Formatting code
black
# Analize code
flake8-django
# Testing code
pytest
pytest-django
# Make random posts for project
Faker
# Give access to get data from frontend
django-cors-headers
# Background processing and cache
celery
redis
django-redis